Use entry points based approach also for build tools
